requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pinecone
sentence-transformers
torch
transformers
mistral_inference
google
google-generativeai
google-genai